model_functions/object_detection/faster_cnn/requirements.txt
numpy
torch==1.4.0+cpu 
torchvision==0.5.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
Pillow==6.1
Flask